:recycle: api: rename createSessionController to sessionController
