+++ api/src/certification/session/routes.js
@@ -1,4 +1,4 @@
-import * as createSession from './application/create-session-route.js';
+import * as session from './application/session-route.js';
 import * as sessionLiveAlert from './application/session-live-alert-route.js';
 import * as attendanceSheet from './application/attendance-sheet-route.js';
 import * as certificationCandidate from './application/certification-candidate-route.js';
@@ -6,7 +6,7 @@ import * as invigilatorKit from './application/invigilator-kit-route.js';
 import * as updateCpfImportStatus from './application/update-cpf-import-status-route.js';
 
 const certificationSessionRoutes = [
-  createSession,
+  session,
   sessionLiveAlert,
   attendanceSheet,
   certificationCandidate,
